Fixed race condition on monaco editor initialization
